+++ Mem1/train/verl/workers/critic/dp_critic.py
@@ -57,7 +57,7 @@ class DataParallelPPOCritic(BasePPOCritic):
             batch, seqlen = input_ids.shape
             attention_mask = micro_batch['attention_mask']
             if 'attention_mask_4d' in micro_batch:
-                attention_mask_4d = micro_batch['attention_mask_4d']
+                attention_mask_4d = micro_batch['attention_mask_4d'].to(torch.bfloat16)
             else:
                 attention_mask_4d = None
             position_ids = micro_batch['position_ids']
